UI Card detail: rename cards module to card module
