+++ ui/main/src/app/modules/card/card.component.scss
