+++ ui/main/src/app/modules/card/components/card-ack/card-ack.component.html
