+++ ui/main/src/app/modules/card/components/card-acks-footer/card-acks-footer.component.ts
