+++ ui/main/src/app/modules/card/components/card-body/card-body.component.scss
