+++ ui/main/src/app/modules/card/components/card-footer-text/card-footer-text.component.scss
