+++ ui/main/src/app/modules/card/components/card-reponse/card-response.component.ts
