+++ ui/main/src/app/modules/card/services/handlebars.service.ts
